requirements.txt
streamlit>=1.51.0
pandas>=2.0.0
plotly>=5.17.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
tqdm>=4.65.0
schedule>=1.2.0
google-cloud-bigquery>=3.10.0
db-dtypes>=1.0.0
watchdog>=3.0.0
openpyxl>=3.1.0
openai>=1.12.0  # For OpenRouter API integration (LLM SQL generation)
requests>=2.31.0  # For Toast API HTTP calls
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyarrow>=12.0.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
